Define search_form further, add get_mosaic to Collection class
