+++ calisphere/institution_views.py
@@ -3,10 +3,10 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.http import Http404
 from . import constants
-from . import views
 from .facet_filter_type import getCollectionData, getRepositoryData
 from .cache_retry import SOLR_select, json_loads_url
-from .search_form import SearchForm
+from . import search_form
+from .collection_views import Collection
 
 import math
 import re
@@ -164,13 +164,13 @@ def institution_view(request,
         elif institution_type == 'campus':
             extra_filter = 'campus_url: "' + institution_url + '"'
 
-        solr_params = views.solr_encode(params, facet_filter_types)
+        solr_params = search_form.solr_encode(params, facet_filter_types)
         if extra_filter:
             solr_params['fq'].append(extra_filter)
         solr_search = SOLR_select(**solr_params)
 
-        facets = views.facet_query(facet_filter_types, params, solr_search,
-                                   extra_filter)
+        facets = search_form.facet_query(facet_filter_types, params, 
+                                         solr_search, extra_filter)
 
         filter_display = {}
         for filter_type in facet_filter_types:
@@ -182,7 +182,7 @@ def institution_view(request,
                 filter_display[display_name] = list(
                     map(filter_transform, params.getlist(param_name)))
 
-        context = SearchForm(params).context
+        context = search_form.search_defaults(params)
         context.update({
             'filters':
             filter_display,
@@ -220,7 +220,7 @@ def institution_view(request,
                 'campus_slug':
                 institution_details.get('slug'),
                 'related_collections':
-                views.get_related_collections(
+                search_form.get_related_collections(
                     params, slug=institution_details.get('slug'))[0],
                 'form_action':
                 reverse(
@@ -242,8 +242,8 @@ def institution_view(request,
                 'uc_institution':
                 uc_institution,
                 'related_collections':
-                views.get_related_collections(params,
-                                              repository_id=institution_id)[0],
+                search_form.get_related_collections(params,
+                                        repository_id=institution_id)[0],
                 'form_action':
                 reverse(
                     'calisphere:repositoryView',
@@ -335,10 +335,11 @@ def institution_view(request,
                     collection_parts[2],
                     collection_parts[1],
                 ))
-            collection_mosaic = views.get_collection_mosaic(
-                collection_data.get('url'))
-            if collection_mosaic:
-                related_collections.append(collection_mosaic)
+            try:
+                related_collections.append(
+                    Collection(collection_data.get('id')).get_mosaic())
+            except Http404:
+                pass
 
         context = {
             'page': page,
